detect and backup missing segments
This is implemented by replacing get_next_offset with get_next_segment,
which is basically an extended version of the original, except that it
returns the first and last segment to fetch, instead of only the first.
This has also moved the responsibility for deadling with record_limit
to get_next_segment, instead of the run function.

This does _not_ detect whether there are data _older_ than the first
stored segment that can potentially be fetched.

fixes #6